+++ miden-crypto/src/hash/algebraic_sponge/poseidon2/test.rs
@@ -95,6 +95,40 @@ fn test_poseidon2_permutation_deterministic() {
 }
 
 #[test]
+fn test_poseidon2_hash_elements_vs_hash_elements_in_domain() {
+    let input16 = [
+        Felt::new_unchecked(1),
+        Felt::new_unchecked(2),
+        Felt::new_unchecked(3),
+        Felt::new_unchecked(4),
+        Felt::new_unchecked(5),
+        Felt::new_unchecked(6),
+        Felt::new_unchecked(7),
+        Felt::new_unchecked(8),
+        Felt::new_unchecked(9),
+        Felt::new_unchecked(10),
+        Felt::new_unchecked(11),
+        Felt::new_unchecked(12),
+        Felt::new_unchecked(13),
+        Felt::new_unchecked(14),
+        Felt::new_unchecked(15),
+        Felt::new_unchecked(16),
+    ];
+
+    // hash_elements and hash_elements_in_domain should be identical if the domain is set to zero.
+    assert_eq!(
+        Poseidon2::hash_elements(&input16),
+        Poseidon2::hash_elements_in_domain(&input16, Felt::ZERO)
+    );
+
+    // With a non-zero domain set in the latter, the results should differ.
+    assert_ne!(
+        Poseidon2::hash_elements(&input16),
+        Poseidon2::hash_elements_in_domain(&input16, Felt::ONE)
+    );
+}
+
+#[test]
 fn test_poseidon2_hasher_vs_hash_elements() {
     let hasher = Poseidon2Hasher::new(Poseidon2Permutation256);